Added hasValidMoves()
Refactored the code that checks for any valid moves in isInCheckmate() to a new function called hasValidMoves() so that I can also use that function in isInStalemate. isInCheckmate() now returns true if the team is in check and has no valid moves, while isInStalemate() will return true if the team is NOT in check but has no valid moves.